+++ src/main/kotlin/com/hvs/kotlinspringplayground/artist/autoconfigure/ArtistAutoConfiguration.kt
@@ -1,12 +1,12 @@
 package com.hvs.kotlinspringplayground.artist.autoconfigure
 
 import com.hvs.kotlinspringplayground.artist.repository.ArtistRepository
-import com.hvs.kotlinspringplayground.artist.service.ArtistService
+import com.hvs.kotlinspringplayground.artist.service.impl.ArtistService
 import com.hvs.kotlinspringplayground.spotify.service.SpotifyService
 import com.hvs.kotlinspringplayground.outbox.autoconfigure.OutboxAutoConfiguration
 import com.hvs.kotlinspringplayground.outbox.service.OutboxService
 import com.hvs.kotlinspringplayground.tidal.service.TidalService
-import com.hvs.kotlinspringplayground.user.service.UserService
+import com.hvs.kotlinspringplayground.user.service.impl.UserService
 import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
