+++ src/main/kotlin/com/hvs/kotlinspringplayground/artist/event/ReleaseEvent.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.node.ObjectNode
 import com.hvs.kotlinspringplayground.outbox.domain.BaseEvent
 
 class ReleaseEvent(id: String, release: ObjectNode): BaseEvent(
-    type = "com.hvs.kotlinspringplayground.tidal",
+    type = "com.hvs.kotlinspringplayground.spotify",
     resultType = "release",
     resultId = id,
     result = release
