+++ src/main/kotlin/com/hvs/kotlinspringplayground/artist/service/ArtistService.kt
@@ -1,85 +1,18 @@
 package com.hvs.kotlinspringplayground.artist.service
 
-
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import com.hvs.kotlinspringplayground.artist.domain.jpa.Artist
 import com.hvs.kotlinspringplayground.artist.dto.ArtistDataDto
-import com.hvs.kotlinspringplayground.artist.event.ReleaseEvent
-import com.hvs.kotlinspringplayground.artist.repository.ArtistRepository
-import com.hvs.kotlinspringplayground.spotify.service.SpotifyService
-import com.hvs.kotlinspringplayground.outbox.service.OutboxService
-import com.hvs.kotlinspringplayground.tidal.client.response.ArtistResponseData.ArtistData
 import com.hvs.kotlinspringplayground.tidal.domain.Album
-import com.hvs.kotlinspringplayground.tidal.service.TidalService
-import com.hvs.kotlinspringplayground.user.service.UserService
-import com.hvs.kotlinspringplayground.user.UserService
-import java.util.UUID
 
-class ArtistService(
-    private val tidalService: TidalService,
-    private val spotifyService: SpotifyService,
-    private val artistRepository: ArtistRepository,
-    private val userService: UserService,
-    private val outboxService: OutboxService,
-) {
+interface ArtistService {
 
-    // TODO: Pattern to use either tidal or spotify, without it being this services concern
-    // TODO: GettingArtistsByName with autocomplete
-    /** Resource Rest with parameter based filtering
-     *              @GetMapping("/user")
-     * public MappingJacksonValue getUser(@RequestParam List<String> fields) {
-     *     User user = // get user data
-     *     SimpleFilterProvider filterProvider = new SimpleFilterProvider();
-     *     filterProvider.addFilter("userFilter", SimpleBeanPropertyFilter.filterOutAllExcept(new HashSet<>(fields)));
-     *
-     *     MappingJacksonValue mapping = new MappingJacksonValue(user);
-     *     mapping.setFilters(filterProvider);
-     *     return mapping;
-     * }
-     */
+    fun getArtistsFromSpotifyByName(artistName: String): ArtistDataDto?
 
-    fun getArtistsFromSpotifyByName(
-        artistName: String
-    ): ArtistDataDto? = spotifyService.getArtistByName(artistName)
-
-    fun storeArtists() {
-        val artistDataResponse = tidalService.getAllArtists()
-        val artists = artistDataResponse.data.map { artist ->
-            ArtistDataDto(
-                streamingId = artist.id,
-                name = artist.attributes.name,
-            )
-        }
-
-        artists.forEach { artist ->
-            artistRepository.save(Artist.from(artist))
-        }
-    }
-
-
+    fun storeArtists()
 
     fun storeSpotifyArtistsForUser(
         username: String,
         artistIdList: List<String>
-    ) = userService.storeArtistsForUser(username, artistIdList)
+    )
 
-    fun getNewAlbumForArtist(artistId: Int): Album? {
-
-        val artist = tidalService.getArtist(artistId = artistId).data.first()
-
-        val artistAlbumIds = tidalService.getAlbumIdListForArtist(artistId = artist.id.toInt())
-
-        val albums = artistAlbumIds.data.mapNotNull { albumId ->
-            tidalService.getAlbum(albumId.id)
-        }
-
-        outboxService.send {
-            ReleaseEvent(
-                id = UUID.randomUUID().toString(),
-                release = jacksonObjectMapper().createObjectNode(),
-            )
-        }
-
-        return albums.maxByOrNull { it.releaseDate }
-    }
+    fun getNewAlbumForArtist(artistId: Int): Album?
 }
