+++ src/main/kotlin/com/hvs/kotlinspringplayground/user/domain/jpa/User.kt
@@ -12,7 +12,6 @@ import org.hibernate.annotations.JdbcTypeCode
 import org.hibernate.type.SqlTypes
 import java.util.UUID
 
-
 @Entity
 @Table(name = "users")
 data class User(
